Class 137 Health Component
In This class I learned how to create Health component I done this by creating the C++ class by Actor Component and attached it to the Tank and tower Blueprint.

I added Maxx helath and Health in the C++ file of health component and declatred max health = to health in the begin play fuction So everytime game starts the health will be full.

I leaned about how to take dame you need to watch it agin it includes that delegate for damage I created the DamgeTaken Function the HealthComponent C++ file and added a lot of pointer in it fro passing data agin watch this class my future self.