Class 141 CustomPlayerController
In this class I learned to create a player controller C++ class and made a blueprint by that class then assigned that playercontroller blueprint to in the gamemode blueprint.

In this player controller I enbled and diabled the input according to a bool in the playercontroller C++ class,

and show the mouse cursor according to he bool but in this class only functionlity is ready rest of the work will be done in next class

and I also set the mouse cursor type in the PlayerController Blueprint I choose a crosshair.